move cfg handling into its own module
